\name{predIntervals-package}
\alias{predIntervals-package}
\alias{predIntervals}
\docType{package}
\title{
predIntervals
}
\description{
Calculation of prediction and precedence intervals for normal and non-normal data.
}
\details{
\tabular{ll}{
Package: \tab predIntervals\cr
Type: \tab Package\cr
Version: \tab 0.0-1\cr
Date: \tab 2010-18-08\cr
License: \tab GPL \cr
}
}
\author{
Daniel Gerhard <gerhard@biostat.uni-hannover.de>
}
\references{
Hahn, GJ and Meeker, WQ (1991): Statistical Intervals. Wiley, New York.

Chakraborti, S., Van der Laan, P., Van de Wiel, M.A. (2004): A class of distribution-free control charts. Applied Statistics 53(3):443-462.

Danziger, L. and Davis, S.A. (1964): Tables of distribution-free tolerance limits. Annals of Mathematical Statistics 35(3):1361-1365.

Odeh, RE (1990): 2-Sided prediction intervals to contain at least k out of m future observations from a normal distribution. Technometric 32(2): 203-216.
}
\keyword{ package }

