\name{show-methods}
\docType{methods}
\alias{show,PredInterval-method}
\alias{show,PrecInterval-method}
\alias{show,NormalPredInterval-method}
\alias{show,NormalPrecInterval-method}
\alias{show,nparPredInterval-method}
\alias{show,nparPrecInterval-method}
\alias{show,PIlm-method}
\title{ Methods for Function show in Package `predIntervals' }
\description{
 Methods for function \code{show} in Package `predIntervals'
}
\keyword{methods}


